+++ src/test/java/OneTimePasswordGeneratorTest.java
@@ -1,5 +1,7 @@
 import org.junit.jupiter.api.Test;
 
+import java.math.BigInteger;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class OneTimePasswordGeneratorTest {
@@ -9,18 +11,18 @@ class OneTimePasswordGeneratorTest {
     @Test
     void generateWithSHA1() {
         OneTimePasswordGenerator generator = new OneTimePasswordGenerator(HMACAlgorithm.SHA1, secret);
-        assertEquals("317963", generator.generate(1));
+        assertEquals("317963", generator.generate(BigInteger.valueOf(1)));
     }
 
     @Test
     void generateWithSHA256() {
         OneTimePasswordGenerator generator = new OneTimePasswordGenerator(HMACAlgorithm.SHA256, secret);
-        assertEquals("604514", generator.generate(1));
+        assertEquals("604514", generator.generate(BigInteger.valueOf(1)));
     }
 
     @Test
     void generateWithSHA512() {
         OneTimePasswordGenerator generator = new OneTimePasswordGenerator(HMACAlgorithm.SHA512, secret);
-        assertEquals("597655", generator.generate(1));
+        assertEquals("597655", generator.generate(BigInteger.valueOf(1)));
     }
 }
